requirements.txt
python-telegram-bot
youtube-dl
pafy
